cyw43439: add support for CYW43439 wifi chip
Added support for the PIO half-duplex SPI interface for the
RP2040 chip to provide read/write functionalities for communicating with
the CYW43439 chip available on the Raspberry Pi Pico W board. Support
is also added in the board main alongside with a `wifi` feature. A minimal
WiFi HIL for the Cypress 43439/4343W chips was added. Both WiFi HIL and
EthernetTap HIL implementation were provided for the RP2040 PIO.